Replace (broken) enforceMQCompliancy on JmsSender with MQSender
Method enforceMQCompliancy on JmsSender in the core module generated
NoSuchMethodException
"com.ibm.mq.jms.MQQueue.setTargetClient(java.lang.Integer) You need to
have ibm mq". This functionality is now implemented as MQSender in the
IBM module.

As FXF 1 and 2 depend on enforceMQCompliancy method in JMSFacade thier
classes would need to be adjusted too, but because FXF 1 and 2 are not
used anymore they are instead deleted.