Move example src/main/webapp/WEB-INF/classes to src/main/resources
This seems to be the right location. The old location will work with
jetty-maven-plugin when no IAF module is added to the dependencies list
otherwise it will not load springCustom.xml anymore. When the Larva
module was added to the dependencies the Ladybug Test Tool didn't work
anymore on ibis4example.ibissource.org.