Support Unicode supplementary characters in replaceNonValidXmlCharacters
Support Unicode supplementary characters (like a smiley) in
replace/stripNonValidXmlCharacters (which is used in ReplacerPipe). In
the old code a Unicode supplementary character like a smiley was seen as
two characters which would both be replaced/stripped. To be backwards
compatible the Unicode supplementary characters are still
replaced/stripped (by one character instead of two) but can be allowed
using allowUnicodeSupplementaryCharacters