Added parameter BROWSER_DATETIME_FORMAT to send dates in ISO format
By default datetime fields are send as a value in milleseconds. This leads to differences in time if the client is in a different timezone than the server. To fix this problem, the datetime fields have to be send in ISO format. This is done by setting the BROWSER_DATETIME_FORMAT to EXTENDED.