Add parameter BROWSER_DATETIME_FORMAT to send dates in ISO format (#155)
By default datetime fields are send as a value in milleseconds. This leads to differences in time if the client is in a different timezone than the server. To fix this problem, the datetime fields have to be send in ISO format. This is done by setting the BROWSER_DATETIME_FORMAT to EXTENDED.

See: https://www.oasis-open.org/committees/download.php/53091/cmis-extension-browser-datetime-format.pdf