Remove autoIncrement=true attribute in DatabaseChangelog.xml (#230)
Removed the autoIncrement=true attribute for creating the tables
IBISDATA and IBISTEMP when using liquibase for Ibis4Test. This is
because this attribute is not necessary and caused an incorrect query
when running Ibis4Test with MySQL and MariaDB. Note that if you have
already run Ibis4Test before using a local database you need to drop
your Databasechangelog table before running Ibis4Test with liquibase
again.