Fix ibisdoc.json
BlobLineIteratingPipe, ClobLineIteratingPipe and ResultSetIteratingPipe
inherit from JdbcIteratingPipeBase. That class has @IbisDocRef
annotations to FixedQuerySender. But that is not the end of the
referencing chain. Property datasourceName actually comes from a
superclass of FixedQuerySender, which is JdbcFacade. The expected value
is changed to include datasourceName. Furthermore, JdbcFacade is added
to the superclasses of the three mentioned classes.