IbisDoc: Continue separating content from information gathering (#1035)
This is the first step to generate the Frank configuration schema, the
FrankDoc and the syntax 1.0 lookup compile-time. The plan is to
separate the logic to gather the relevant information from the
logic that writes the generated files.

BlobLineIteratingPipe, ClobLineIteratingPipe and ResultSetIteratingPipe
inherit from JdbcIteratingPipeBase. That class has @IbisDocRef
annotations to FixedQuerySender. But that is not the end of the
referencing chain. Property datasourceName actually comes from a
superclass of FixedQuerySender, which is JdbcFacade. The expected value
is changed to include datasourceName. Furthermore, JdbcFacade is added
to the superclasses of the three mentioned classes.