Rename FRANKDOC.md because it is not what we want
I want to close this pull request without losing the work I did. In the
future I may rewrite some of the text. We need a guide on how to develop
the Frank!Framework. That is the place to explain how Java classes and
Java methods should be organized and named. That explanation should not
be combined with explaining the Frank!Doc like I was trying in this PR.