Ensure all CXF endpoints are loaded as Servlet instead of a Filter (#6871)
Since Spring Boot 3.x all filter beans are picked up and loaded first, and for some reason the magnificent CXF people decided that their 'CXFServlet' class can sercretly behave as a filter as well.
This breaks the load order as well the way initParameters are loaded, so for us this is a big no-go. Wrapping it in a class that only implements servlet (and not filter) ensures correct loading behavior.