Move placeholder text on edit-trigger page to a 'hint' (#8032)
It's a lot of text, and it overflows if there is not enough space. It's not the best solution but for now it's the easiest...
Placeholders do not work when there is a default value. Though this not the best solution, and perhaps we should revamp all the forms, but at least it improves the current user-flow a bit...